ip: Fix recvmsg with ancillary data
A successful call to `recvmsg()` causes `msg.msg_controllen` to contain the length of the received ancillary data. Quoting the manpage for `recvmsg`:
"The field msg_control, which has length msg_controllen, ...; upon return from a successful call it will contain the length of the control message sequence."

Therefore, we must reset this value after each call to `recvmsg()`.

This PR fixes `ip monitor` running with the `all-nsid` option - When passing `all-nsid` the kernel passes the nsid as ancillary data.
However, if while `ip monitor` runs an event on the current netns is received, then no ancillary data will be sent, causing the `msg.msg_controllen` to be set to 0.

This means future calls to `recvmsg()`, which otherwise would receive the nsid as ancillary data, will instead have `MSG_CTRUNC` being set in `msg.msg_flags` (and `[nsid current]` will be printed for all eternity, which is fixable only by stopping and running `ip monitor` again)